hoc-sys/loader/config-pc/config-device - move unsafe freqs setting to hoc-sys
This allows stuff to be changed on the fly

+++ Source/sys-clk/common/include/sysclk/clock_manager.h
@@ -30,8 +30,8 @@ typedef struct
 typedef struct
 {
     union {
-        uint32_t mhz[SysClkProfile_EnumMax * SysClkModule_EnumMax];
-        uint32_t mhzMap[SysClkProfile_EnumMax][SysClkModule_EnumMax];
+        uint32_t mhz[(uint32_t)SysClkProfile_EnumMax * (uint32_t)SysClkModule_EnumMax];
+        uint32_t mhzMap[(uint32_t)SysClkProfile_EnumMax][(uint32_t)SysClkModule_EnumMax];
     };
 } SysClkTitleProfileList;
